# docs/swagger.yaml
basePath: /swagger/authz-service
definitions:
  common.Page:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  model.Resource:
    properties:
      api_url:
        description: api鉴权用到
        type: string
      data_conditions:
        description: 数据条件（数据鉴权用到，存放对象的json字符串)
        type: string
      id:
        description: 唯一标识,例如UI-001-001
        type: string
      module:
        description: 模块
        type: string
      name:
        description: 名称
        type: string
      parent_id:
        description: parent_id
        type: string
      service_name:
        description: 服务名
        type: string
      service_name_cn:
        description: 中文服务名
        type: string
      sort_index:
        description: sort_index
        type: integer
      support_ops:
        description: 支持的操作(字典值,1,2,3)
        type: string
      type:
        description: 分类（1:菜单,2:api,3:功能点,4:数据)
        type: integer
    type: object
...
info:
  contact: {}
  description: authz-service API
  title: authz-service API
  version: "1.0"
paths:
...
  /resource:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: parent_id
        in: query
        name: parent_id
        type: string
      - description: module
        in: query
        name: module
        type: string
      - description: service_name
        in: query
        name: service_name
        type: string
      - description: service_name_cn
        in: query
        name: service_name_cn
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: api_url
        in: query
        name: api_url
        type: string
      - description: sort_index
        in: query
        name: sort_index
        type: string
      - description: data_conditions
        in: query
        name: data_conditions
        type: string
      - description: support_ops
        in: query
        name: support_ops
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Resource'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - Resource
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Resource'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Resource'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - Resource
  /resource/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Resource'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - Resource
  /resource/batch-delete:
    post:
      consumes:
      - application/json
      description: batch delete
      parameters:
      - description: id array
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch delete
      tags:
      - Resource
  /resource/batch-upsert:
    post:
      consumes:
      - application/json
      description: batch update
      parameters:
      - description: objects array
        in: body
        name: entities
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch update
      tags:
      - Resource
  /resource/groupby:
    get:
      description: GroupBy, for example,  _select=level, then return  {level_val1:sum1,level_val2:sum2},
        _where can input status=0
      parameters:
      - description: _select
        in: query
        name: _select
        required: true
        type: string
      - description: _where
        in: query
        name: _where
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: GroupBy
      tags:
      - Resource
  /resource/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: parent_id
        in: query
        name: parent_id
        type: string
      - description: module
        in: query
        name: module
        type: string
      - description: service_name
        in: query
        name: service_name
        type: string
      - description: service_name_cn
        in: query
        name: service_name_cn
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: api_url
        in: query
        name: api_url
        type: string
      - description: sort_index
        in: query
        name: sort_index
        type: string
      - description: data_conditions
        in: query
        name: data_conditions
        type: string
      - description: support_ops
        in: query
        name: support_ops
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.Resource'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - Resource
...
swagger: "2.0"